Added functions for local cache modal
